requirements.txt
flask 
pillow 
tweepy